requirements.txt
watchdog
pandas
python-dotenv
pydantic
python-dateutil
tqdm
azure-ai-formrecognizer
reportlab
Pillow
pdf2image
openai
pdfminer.six
fastapi
itsdangerous
uvicorn
jinja2
python-multipart
pydantic-settings
passlib[bcrypt]
pytest
lxml
PyYAML
